Split conformance tests for deduplication duration validation and include on all valid ledgers.
