+++ ledger/daml-on-sql/BUILD.bazel
@@ -115,6 +115,7 @@ conformance_test(
         "--verbose",
         "--additional=MultiPartySubmissionIT",
         "--additional=CommandDeduplicationParallelIT",
+        "--additional=CommandDeduplicationPeriodValidationIT",
         "--additional=CompletionDeduplicationInfoITCommandService",
         "--additional=CompletionDeduplicationInfoITCommandSubmissionService",
         "--additional=ContractIdIT:Accept",
