+++ ledger/ledger-on-memory/BUILD.bazel
@@ -153,6 +153,7 @@ conformance_test(
         # Notoriously times-out
         #"--additional=ParticipantPruningIT",
         "--additional=CommandDeduplicationParallelIT",
+        "--additional=CommandDeduplicationPeriodValidationIT",
         "--additional=CompletionDeduplicationInfoITCommandService",
         "--additional=CompletionDeduplicationInfoITCommandSubmissionService",
         "--additional=KVCommandDeduplicationIT",
@@ -180,6 +181,7 @@ conformance_test(
         # Notoriously times-out
         #"--additional=ParticipantPruningIT",
         "--additional=CommandDeduplicationParallelIT",
+        "--additional=CommandDeduplicationPeriodValidationIT",
         "--additional=CompletionDeduplicationInfoITCommandService",
         "--additional=CompletionDeduplicationInfoITCommandSubmissionService",
         "--additional=KVCommandDeduplicationIT",
@@ -210,6 +212,7 @@ conformance_test(
         "--additional=KVCommandDeduplicationIT",
         # The following two tests don't actually care about multi-participant but they do need append-only.
         "--additional=CommandDeduplicationParallelIT",
+        "--additional=CommandDeduplicationPeriodValidationIT",
         "--additional=CompletionDeduplicationInfoITCommandService",
         "--additional=CompletionDeduplicationInfoITCommandSubmissionService",
         "--exclude=CommandDeduplicationIT",  # It's a KV ledger so it needs the KV variant
@@ -235,6 +238,7 @@ conformance_test(
         # Notoriously times-out
         #"--additional=ParticipantPruningIT",
         "--additional=CommandDeduplicationParallelIT",
+        "--additional=CommandDeduplicationPeriodValidationIT",
         "--additional=CompletionDeduplicationInfoITCommandService",
         "--additional=CompletionDeduplicationInfoITCommandSubmissionService",
         "--additional=KVCommandDeduplicationIT",
