+++ ledger/sandbox/BUILD.bazel
@@ -274,6 +274,7 @@ conformance_test(
     test_tool_args = [
         "--open-world",
         "--additional=CommandDeduplicationParallelIT",
+        "--additional=CommandDeduplicationPeriodValidationIT",
         "--additional=CompletionDeduplicationInfoITCommandService",
         "--additional=CompletionDeduplicationInfoITCommandSubmissionService",
         "--additional=KVCommandDeduplicationIT",
@@ -311,6 +312,7 @@ conformance_test(
         "--static-time",
         "--open-world",
         "--additional=CommandDeduplicationParallelIT",
+        "--additional=CommandDeduplicationPeriodValidationIT",
         "--additional=CompletionDeduplicationInfoITCommandService",
         "--additional=CompletionDeduplicationInfoITCommandSubmissionService",
         "--additional=KVCommandDeduplicationIT",
